chore: switch to use SASS for AutolinkHeader component
This demonstrates the value of only importing the necessary variables
into sass files. Files take much less time to compile during dev.

Also shown here is the usage of sass modules to ensure no there are
no name collisions, explicit dependencies and no global scope.
