+++ packages/gatsby-theme-carbon/src/components/AutolinkHeader/index.js
@@ -1,0 +1,3 @@
+import AutolinkHeader from './AutolinkHeader';
+
+export default AutolinkHeader;
